+++ app/src/main/java/eu/kanade/tachiyomi/data/track/bangumi/BangumiApi.kt
@@ -115,7 +115,7 @@ class BangumiApi(
 
                     track.status = it.status?.id?.toInt() ?: Bangumi.DEFAULT_STATUS
                     track.last_chapter_read = it.epStatus!!.toFloat()
-                    track.score = it.rating!!.toFloat()
+                    track.score = it.rating!!.toDouble()
                     track
                 }
         }
