+++ app/src/main/java/eu/kanade/tachiyomi/data/track/kitsu/dto/KitsuListSearch.kt
@@ -40,7 +40,7 @@ data class KitsuListSearchResult(
                 "planned" -> Kitsu.PLAN_TO_READ
                 else -> throw Exception("Unknown status")
             }
-            score = (userDataAttrs.ratingTwenty?.let { it / 2.0 } ?: 0.0).toFloat()
+            score = (userDataAttrs.ratingTwenty?.let { it / 2.0 } ?: 0.0)
             last_chapter_read = userDataAttrs.progress.toFloat()
         }
     }
