+++ app/src/main/java/eu/kanade/tachiyomi/data/track/shikimori/Shikimori.kt
@@ -83,7 +83,7 @@ class Shikimori(private val context: Context, id: Long) : TrackService(id) {
     }
 
     override suspend fun add(track: Track): Track {
-        track.score = DEFAULT_SCORE.toFloat()
+        track.score = DEFAULT_SCORE.toDouble()
         track.status = DEFAULT_STATUS
         updateNewTrackInfo(track)
         return api.addLibManga(track, getUsername())
