+++ app/src/main/java/eu/kanade/tachiyomi/ui/more/stats/StatsPresenter.kt
@@ -80,7 +80,7 @@ class StatsPresenter(
         return downloadManager.getDownloadCount(manga.manga)
     }
 
-    fun get10PointScore(track: Track): Float? {
+    fun get10PointScore(track: Track): Double? {
         val service = trackManager.getService(track.sync_id)
         return service?.get10PointScore(track.score)
     }
